account/api/versioned/v1/forex/views.py
from rest_framework import viewsets

from api.bases.forex.models import ExchangeRate
from .serializers import ExchangeRateSerializer


class ExchangeRateViewSet(viewsets.ModelViewSet):
    queryset = ExchangeRate.objects.all()
    serializer_class = ExchangeRateSerializer
