account/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
Django==3.0.3
django-celery-beat==2.0.0
django-celery-results==2.0.0
django-cors-headers==3.2.1
django-fernet-fields==0.6
django-filter==2.2.0
django-model-utils==4.0.0
django-redis==4.12.1
django-reversion==3.0.7
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.1.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
elastic-apm==4.2.2
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==2.0.1
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
aiohttp==3.7.4.post0
django-admin-inline-paginator==0.3.0